fix: conversion bug preventing server form being able to process client messages

+++ Sources/tss-client-swift/TSSMessages/PrecomputeStatus.swift
@@ -2,5 +2,5 @@ import Foundation
 
 public struct TssPrecomputeUpdate : Codable {
     let session: String
-    let party: Int64
+    let party: Int
 }
