+++ Sources/tss-client-swift/TSSMessages/RecvMsg.swift
@@ -3,8 +3,8 @@ import SocketIO
 
 public struct TssRecvMsg : Codable {
     let session: String
-    let sender: Int64
-    let recipient: Int64
+    let sender: Int
+    let recipient: Int
     let msg_type: String
     let msg_data: String
 }
